+++ src/main/java/com/team2final/minglecrm/controller/customer/CustomerController.java
@@ -1,67 +1,51 @@
-    package com.team2final.minglecrm.controller.customer;
+package com.team2final.minglecrm.controller.customer;
 
-    import com.team2final.minglecrm.dto.CustomerListResponseDTO;
-    import com.team2final.minglecrm.dto.CustomerUpdateDTO;
-    import com.team2final.minglecrm.service.customer.CustomerService;
-    import lombok.RequiredArgsConstructor;
-    import org.springframework.http.ResponseEntity;
-    import org.springframework.web.bind.annotation.*;
+import com.team2final.minglecrm.controller.customer.request.CustomerUpdateRequest;
+import com.team2final.minglecrm.controller.customer.response.CustomerListResponse;
+import com.team2final.minglecrm.service.customer.CustomerService;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-    import java.util.List;
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/customers")
+public class CustomerController {
 
-    @RestController
-    @RequiredArgsConstructor
-    @RequestMapping("/api/v1/customers")
-    public class CustomerController {
+    private final CustomerService customerService;
 
-        private final CustomerService customerServiceHP;
-
-        // 토큰 있다는 가정
-//        private final JwtService jwtService;
-
-
-        @GetMapping("")
-        public ResponseEntity<List<CustomerListResponseDTO>> getAllCustomers() {
-//            String token = jwtService.extractTokenFromRequest(request);
-//            String userEmail = jwtService.extractUserName(token);
-//            User user = userService.findByEmail(userEmail);
-//
-//            if(!user.isAdminYn()) {
-//                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//            }
-            List<CustomerListResponseDTO> customers = customerServiceHP.getAllCustomer();
-            return ResponseEntity.ok(customers);
-        }
-
-        @GetMapping("/username/{username}")
-        public ResponseEntity<List<CustomerListResponseDTO>> getCustomersByUserId(@PathVariable("username") String userName) {
-//            String token = jwtService.extractTokenFromRequest(request);
-//            String userEmail = jwtService.extractUserName(token);
-//            User user = userService.findByEmail(userEmail);
-//
-//            if(!user.isAdminYn()) {
-//                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//            }
-            List<CustomerListResponseDTO> customersByUserId = customerServiceHP.getCustomersByUserId(userName);
-            return ResponseEntity.ok(customersByUserId);
-        }
-
-        @GetMapping("/group/{customerGroup}")
-        public ResponseEntity<List<CustomerListResponseDTO>> getCustomersByCustomerGroup(@PathVariable("customerGroup") String customerGroup) {
-//            String token = jwtService.extractTokenFromRequest(request);
-//            String userEmail = jwtService.extractUserName(token);
-//            User user = userService.findByEmail(userEmail);
-//
-//            if(!user.isAdminYn()) {
-//                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//            }
-            List<CustomerListResponseDTO> customersByGroup = customerServiceHP.getCustomersByCustomerGroup(customerGroup);
-            return ResponseEntity.ok(customersByGroup);
-        }
-
-        @PatchMapping("/{customerId}")
-        public ResponseEntity<Void> updateCustomer(@PathVariable ("customerId") Long Id, @RequestBody CustomerUpdateDTO customerUpdateDTO) {
-            customerServiceHP.updateCustomer(Id, customerUpdateDTO);
-            return ResponseEntity.ok().build();
-        }
+    @GetMapping("")
+    public ResponseEntity<List<CustomerListResponse>> getAllCustomers() {
+        List<CustomerListResponse> customers = customerService.getAllCustomer();
+        return ResponseEntity.ok(customers);
     }
+
+    @GetMapping("/employee/{employee}")
+    public ResponseEntity<List<CustomerListResponse>> getCustomersByUserId(
+            @PathVariable("employee") String employeeName) {
+        List<CustomerListResponse> customersByUserId = customerService.getCustomersByUserId(
+                employeeName);
+        return ResponseEntity.ok(customersByUserId);
+    }
+
+    @GetMapping("/group/{customerGroup}")
+    public ResponseEntity<List<CustomerListResponse>> getCustomersByCustomerGroup(
+            @PathVariable("customerGroup") String customerGroup) {
+        List<CustomerListResponse> customersByGroup = customerService.getCustomersByCustomerGroup(
+                customerGroup);
+        return ResponseEntity.ok(customersByGroup);
+    }
+
+    @PatchMapping("/{customerId}")
+    public ResponseEntity<Void> updateCustomer(@PathVariable("customerId") Long Id,
+            @RequestBody CustomerUpdateRequest customerUpdateRequest) {
+        customerService.updateCustomer(Id, customerUpdateRequest);
+        return ResponseEntity.ok().build();
+    }
+}
