+++ src/main/java/com/team2final/minglecrm/controller/employee/request/SignInRequest.java
@@ -1,0 +1,11 @@
+package com.team2final.minglecrm.controller.employee.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SignInRequest {
+    private String email;
+    private String password;
+}
