+++ src/main/java/com/team2final/minglecrm/controller/employee/request/SignUpRequest.java
@@ -1,0 +1,20 @@
+package com.team2final.minglecrm.controller.employee.request;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class SignUpRequest {
+
+    private String email;
+    private String name;
+    private String password;
+
+    @Builder
+    public SignUpRequest(String email, String name, String password) {
+        this.email = email;
+        this.name = name;
+        this.password = password;
+    }
+}
