--- src/main/java/com/team2final/minglecrm/controller/user/dto/SignInRequestDTO.java
@@ -1,11 +1,0 @@
-package com.team2final.minglecrm.controller.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SignInRequestDTO {
-    private String email;
-    private String password;
-}
