--- src/main/java/com/team2final/minglecrm/controller/user/dto/SignUpRequestDTO.java
@@ -1,20 +1,0 @@
-package com.team2final.minglecrm.controller.user.dto;
-
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class SignUpRequestDTO {
-
-    private String email;
-    private String name;
-    private String password;
-
-    @Builder
-    public SignUpRequestDTO(String email, String name, String password) {
-        this.email = email;
-        this.name = name;
-        this.password = password;
-    }
-}
