+++ src/main/java/com/team2final/minglecrm/entity/BaseTimeEntity.java
@@ -1,0 +1,27 @@
+package com.team2final.minglecrm.entity;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+
+/**
+ * EntityListeners 참고 : https://mopil.tistory.com/46
+ *
+ */
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseTimeEntity {
+
+    @CreatedDate
+    private LocalDateTime registerDate;
+
+    @LastModifiedDate
+    private LocalDateTime lastConsulationDate;
+}
