+++ src/main/java/com/team2final/minglecrm/persistence/repository/employee/EmployeeRepository.java
@@ -1,0 +1,16 @@
+package com.team2final.minglecrm.persistence.repository.employee;
+
+import com.team2final.minglecrm.entity.employee.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
+    boolean existsByEmail(String email);
+
+    Optional<Employee> findByEmail(String email);
+
+}
