--- src/main/java/com/team2final/minglecrm/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.team2final.minglecrm.repository;
-
-import com.team2final.minglecrm.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-
-    Optional<User> findByEmail(String email);
-
-}
