--- src/main/java/com/team2final/minglecrm/service/user/UserService.java
@@ -1,21 +1,0 @@
-package com.team2final.minglecrm.service.user;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.team2final.minglecrm.controller.user.dto.SignInRequestDTO;
-import com.team2final.minglecrm.controller.user.dto.SignInResponseDTO;
-import com.team2final.minglecrm.controller.user.dto.SignUpRequestDTO;
-import com.team2final.minglecrm.controller.user.dto.SignUpResponseDTO;
-import org.springframework.transaction.annotation.Transactional;
-
-
-public interface UserService {
-
-    @Transactional
-    SignUpResponseDTO signUp(SignUpRequestDTO signUpRequestDTO);
-
-    @Transactional
-    SignInResponseDTO signIn(SignInRequestDTO signInRequestDTO);
-
-    @Transactional
-    Void logout(String atk) throws JsonProcessingException;
-}
